requirements.txt
viam-sdk==0.36.0
typing-extensions
